+++ src/test/java/CheckingTest.java
@@ -43,7 +43,7 @@ public class CheckingTest {
         return account.getID();
     }
 
-    private int accountBalance(Account account) {
+    private double accountBalance(Account account) {
         return account.getBalance();
     }
 }
